Move conversation kill switch updates to WorkspaceResource
